Add files via upload
Updated TuCaurto Security as of 19-05

+++ TuCaurto-SecirtyService/src/main/java/in/tucaurto/dao/UserLoginRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.repository.CrudRepository;
 
 import in.tucaurto.entity.UserLogin;
 
-public interface UserLoginRepository extends CrudRepository<UserLogin, Integer> {
+public interface UserLoginRepository extends CrudRepository<UserLogin, Long> {
     UserLogin findByUsername(String username);
 }
