+++ TuCaurto-SecirtyService/src/main/java/in/tucaurto/dao/UserRepository.java
@@ -4,7 +4,7 @@ import org.springframework.data.repository.CrudRepository;
 
 import in.tucaurto.entity.User;
 
-public interface UserRepository extends CrudRepository<User, Integer>
+public interface UserRepository extends CrudRepository<User, Long>
 {
 	
 }
